LogicSettings: usePages hook
