+++ front/app/components/FormBuilder/components/FormBuilderSettings/LogicSettings/index.tsx
@@ -4,13 +4,9 @@ import { Box, colors } from '@citizenlab/cl2-component-library';
 import { get } from 'lodash-es';
 import { useFormContext } from 'react-hook-form';
 
-import {
-  IFlatCustomField,
-  IFlatCustomFieldWithIndex,
-} from 'api/custom_fields/types';
+import { IFlatCustomFieldWithIndex } from 'api/custom_fields/types';
 
 import useLocale from 'hooks/useLocale';
-import useLocalize from 'hooks/useLocalize';
 
 import { FormBuilderConfig } from 'components/FormBuilder/utils';
 import Warning from 'components/UI/Warning';
@@ -19,10 +15,10 @@ import { useIntl, FormattedMessage } from 'utils/cl-intl';
 import { isNilOrError } from 'utils/helperUtils';
 
 import messages from '../../messages';
-import { getFieldNumbers } from '../../utils';
 
 import { PageRuleInput } from './PageRuleInput';
 import { QuestionRuleInput } from './QuestionRuleInput';
+import usePages from './usePages';
 
 export type PageListType =
   | {
@@ -50,14 +46,12 @@ const LogicSettings = ({ field, builderConfig }: LogicSettingsProps) => {
     formState: { errors: formContextErrors },
   } = useFormContext();
   const locale = useLocale();
-  const localize = useLocalize();
-  const formCustomFields: IFlatCustomField[] = watch('customFields');
   const selectOptions = watch(`customFields.${field.index}.options`);
   const linearScaleMaximum = watch(`customFields.${field.index}.maximum`);
   const fieldRequired = watch(`customFields.${field.index}.required`);
-  const fieldType = watch(`customFields.${field.index}.input_type`);
   const error = get(formContextErrors, `customFields.${field.index}.logic`);
   const validationError = error?.message as string | undefined;
+  const pages = usePages(field);
 
   // For Select Field
   let answers: AnswersType = selectOptions
@@ -95,66 +89,6 @@ const LogicSettings = ({ field, builderConfig }: LogicSettingsProps) => {
       });
     }
   }
-
-  const getPageList = () => {
-    const fieldNumbers = getFieldNumbers(formCustomFields);
-    const pageArray: { value: string; label: string }[] = [];
-
-    formCustomFields.forEach((field, i) => {
-      if (field.input_type === 'page') {
-        const isLastPage = i === formCustomFields.length - 1;
-
-        const pageTitle = localize(field.title_multiloc);
-        const pageLabel = isLastPage
-          ? formatMessage(messages.lastPage)
-          : `${formatMessage(messages.page)} ${fieldNumbers[field.id]}${
-              pageTitle
-                ? `: ${
-                    pageTitle.length > 25
-                      ? `${pageTitle.slice(0, 25)}...`
-                      : pageTitle
-                  }`
-                : ''
-            }`;
-
-        pageArray.push({
-          value: field.temp_id || field.id,
-          label: pageLabel,
-        });
-      }
-    });
-    return pageArray;
-  };
-
-  // Which page is the current question on?
-  // Technically there should always be a current page ID and null should never be returned
-  const getCurrentPageId = (
-    field: IFlatCustomFieldWithIndex
-  ): string | null => {
-    if (fieldType === 'page') return field.id;
-
-    let pageId: string | null = null;
-    for (const formCustomField of formCustomFields) {
-      if (formCustomField.input_type === 'page') pageId = formCustomField.id;
-      if (formCustomField.id === field.id) return pageId;
-    }
-    return null;
-  };
-
-  const pageOptions = getPageList();
-
-  // Current and previous pages should be disabled in select options
-  let disablePage = true;
-  const pages: PageListType = pageOptions.map((page) => {
-    const newPage = {
-      ...page,
-      disabled: disablePage,
-    };
-    if (page.value === getCurrentPageId(field)) {
-      disablePage = false;
-    }
-    return newPage;
-  });
 
   return (
     <>
